Revert "Feature/kopiere skjema ved ny behandling (#2460)"
This reverts commit ffe1988068db4bd83a439349ea925fbc821ae28f.
