+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/eøs/kompetanse/KompetanseService.kt
@@ -45,8 +45,4 @@ class KompetanseService(
 
         serviceDelegate.lagreSkjemaDifferanse(gjeldendeKompetanser, oppdaterteKompetanser)
     }
-
-    @Transactional
-    fun kopierOgErstattKompetanser(fraBehandlingId: Long, tilBehandlingId: Long) =
-        serviceDelegate.kopierOgErstattSkjemaer(fraBehandlingId, tilBehandlingId)
 }
