+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/eøs/valutakurs/ValutakursService.kt
@@ -3,7 +3,6 @@ package no.nav.familie.ba.sak.kjerne.eøs.valutakurs
 import no.nav.familie.ba.sak.kjerne.eøs.felles.PeriodeOgBarnSkjemaService
 import no.nav.familie.ba.sak.kjerne.steg.TilbakestillBehandlingService
 import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
 
 @Service
 class ValutakursService(
@@ -20,8 +19,4 @@ class ValutakursService(
 
     fun slettValutakurs(valutakursId: Long) =
         serviceDelegate.slettSkjema(valutakursId)
-
-    @Transactional
-    fun kopierOgErstattValutakurser(fraBehandlingId: Long, tilBehandlingId: Long) =
-        serviceDelegate.kopierOgErstattSkjemaer(fraBehandlingId, tilBehandlingId)
 }
