+++ src/test/integrasjonstester/kotlin/no/nav/familie/ba/sak/kjerne/vilkårsvurdering/VilkårsvurderingFlyttResultaterTest.kt
@@ -142,7 +142,7 @@ class VilkårsvurderingFlyttResultaterTest(
 
         // Lager revurdering når utvidet ikke løper, så underkategorien er ordinær
         kjørStegprosessForBehandling(
-            tilSteg = StegType.REGISTRERE_GRUNNLAG_FOR_NY_BEHANDLING,
+            tilSteg = StegType.REGISTRERE_PERSONGRUNNLAG,
             søkerFnr = søker,
             barnasIdenter = listOf(barn1, barn2),
             underkategori = BehandlingUnderkategori.ORDINÆR,
